requirements.txt
Flask==3.0.0
gunicorn==21.2.0
langchain==0.1.0
langchain-google-genai==0.0.5
langchain-community==0.0.10
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pypdf==3.17.4
python-dotenv==1.0.0